pyproject.toml
[project]
name = "photos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geopy>=2.4.1",
    "ipython>=9.3.0",
    "osxphotos>=0.72.1",
    "pillow>=11.2.1",
]
